lab6/requirements.txt
tensorflow
matplotlib
pillow
ipykernel
